+++ bin/cide.conf
@@ -30,5 +30,5 @@ tools.staticdir.dir = "static"
 # WebSocket
 [/ws]
 tools.websocket.on = True
-tools.websocket.handler_cls = cide.server.index.Publisher
+tools.websocket.handler_cls = cide.server.ideController.IDEWebSocket
